+++ app/ix/core/search/text/TextIndexer.java
@@ -15,20 +15,7 @@ import java.io.Serializable;
 import java.nio.file.Files;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Set;
-import java.util.TreeMap;
+import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.ExecutorService;
@@ -893,9 +880,12 @@ public class TextIndexer implements Closeable, ReIndexListener {
 
 		List<SuggestResult> suggest(CharSequence key, int max) throws IOException {
 			refreshIfDirty();
-			return lookup.lookup(key, null, false, max).stream()
+            return lookup.lookup(key, null, false, max).stream()
+//                    .peek(l -> System.out.println(l.key))
 					.map(r -> new SuggestResult(r.payload.utf8ToString(), r.key, r.value))
 					.collect(Collectors.toList());
+
+
 		}
 	}